move the validation method as private to git package
