+++ pkg/mod/index.go
@@ -18,7 +18,6 @@ package mod
 
 import (
 	"bufio"
-	"fmt"
 	"io"
 	"path/filepath"
 	"slices"
@@ -198,35 +197,6 @@ func (w WorkDir) AddTarget(target string, folderName string, dep KloneSource) er
 
 func cleanRelativePath(src string) string {
 	return filepath.Join(".", filepath.Clean(filepath.Join("/", src)))
-}
-
-// ValidateRepoURL rejects repo_url values that could be re-interpreted by git
-// as a command-line option or as a helper transport (e.g. ext::sh -c …).
-func ValidateRepoURL(repoURL string) error {
-	if repoURL == "" {
-		return fmt.Errorf("repo_url is empty")
-	}
-	if strings.HasPrefix(repoURL, "-") {
-		return fmt.Errorf("repo_url %q starts with '-', refusing to pass it to git", repoURL)
-	}
-	if strings.Contains(repoURL, "::") {
-		return fmt.Errorf("repo_url %q uses a helper transport ('::'), which is not allowed", repoURL)
-	}
-	switch {
-	case strings.HasPrefix(repoURL, "https://"),
-		strings.HasPrefix(repoURL, "http://"),
-		strings.HasPrefix(repoURL, "ssh://"),
-		strings.HasPrefix(repoURL, "git://"),
-		strings.HasPrefix(repoURL, "file://"),
-		strings.HasPrefix(repoURL, "/"):
-		return nil
-	}
-	if at := strings.Index(repoURL, "@"); at > 0 {
-		if colon := strings.Index(repoURL, ":"); colon > at {
-			return nil
-		}
-	}
-	return fmt.Errorf("repo_url %q does not use an allowed scheme (https/http/ssh/git/file/local path/scp-like)", repoURL)
 }
 
 func (w WorkDir) FetchTargets(
